Move ignore warnings to separate module
This function is imported by the style
API, so that depends on Flask. Move
the function to its own module to
avoid bringing in the webserver
stuff when importing the style API.

Refs #1590
