+++ datacube_ows/warnings.py
@@ -1,0 +1,16 @@
+# This file is part of datacube-ows, part of the Open Data Cube project.
+# See https://opendatacube.org for more information.
+#
+# Copyright (c) 2017-2024 OWS Contributors
+# SPDX-License-Identifier: Apache-2.0
+from __future__ import annotations
+
+import warnings
+
+
+def initialise_ignorable_warnings() -> None:
+    # Suppress annoying rasterio warning message every time we write to
+    # a non-georeferenced image format.
+    from rasterio.errors import NotGeoreferencedWarning
+
+    warnings.simplefilter("ignore", category=NotGeoreferencedWarning)
